+++ src/oid4vp/verifier.rs
@@ -3,9 +3,8 @@
 mod dcql;
 mod metadata;
 mod request;
-mod response;
 
 pub use dcql::*;
 pub use metadata::*;
 pub use request::*;
-pub use response::*;
+
